tracing: 新しいトレース機能を追加し、古い関数を非推奨に設定
